Add Board#allowed_moves(piece)
It is better if the board can tell us all the allowed moves
of a piece, instead of querying it for each possibility. That
is nicer to test, since we want our test against all squares.
Note that there was a bug in the code, that was uncovered by
this test.

It will also be helpful later, when we develop a UI.